% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRecords.R
\name{matchRecords}
\alias{matchRecords}
\title{matchRecords}
\usage{
matchRecords(
  nuseds.rds.file,
  nuseds.pop.info,
  epad.rds.file,
  mrp.rds.file,
  mrp.rds.details.file,
  out.folder = "OUTPUT",
  tracking.folder = "DATA/TrackingFiles",
  lookup.folder = "DATA/LookupFiles",
  processed.folder = "DATA/LargeFiles_Processed"
)
}
\arguments{
\item{nuseds.rds.file}{path to the nuSEDS input file (in RDS format)}

\item{nuseds.pop.info}{path to the pop/site info file from nuSEDS (in csv format)}

\item{epad.rds.file}{path to the EPAD input file (in RDS format)}

\item{mrp.rds.file}{path to the MRP summary input file (in RDS format)}

\item{out.folder}{path for storing output (CU summary stored there)}

\item{tracking.folder}{path for storing tracking files (site mismatches etc. stored there)}

\item{lookup.folder}{path for storing generated lookup files ("rosetta" files stored there)}

\item{processed.folder}{folder path for storing modified csv files (MRP summaries with CU match stored there)}

\item{epad.rds.details.file}{path to the EPAD details input file (in RDS format)}
}
\description{
Function to match records across databases. Generates various outputs
(rosetta files, mismatch tracking). Target folders are generated if they don't exist
}
\examples{
\dontrun{mergeFiles(nuseds.rds.file = "DATA/RDSFiles/nuSEDS.RDS",
            nuseds.pop.info ="DATA/LookupFiles/aopen-data-portaljanuary-2019conservation_unit_system_sites.csv",
           epad.rds.file =  "DATA/RDSFiles/EPAD.RDS",
           mrp.rds.file = "DATA/RDSFiles/MRP.RDS" )}
}
\keyword{matching/merging}
\keyword{record}
