% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFiles.R
\name{mergeFiles}
\alias{mergeFiles}
\title{mergeFiles}
\usage{
mergeFiles(nuseds.rds.file, nuseds.pop.info, epad.rds.file,
  out.folder = "OUTPUT", tracking.folder = "DATA/TrackingFiles",
  lookup.folder = "DATA/LookupFiles")
}
\arguments{
\item{nuseds.rds.file}{path to the nuSEDS input file (in RDS format)}

\item{nuseds.pop.info}{path to the pop/site info file from nuSEDS (in csv format)}

\item{epad.rds.file}{path to the EPAD input file (in RDS format)}

\item{out.folder}{path for storing output (CU summary stored there)}

\item{tracking.folder}{path for storing tracking files (site mismatches etc. stored there)}

\item{lookup.folder}{path for storing generated lookup files ("rosetta" files stored there)}
}
\description{
Function to merge/match records across databases. Generates various outputs
(rosetta files, mismatch tracking). Target folders are generated if they don't exist
}
\examples{
\dontrun{mergeFiles(nuseds.rds.file = "DATA/RDSFiles/nuSEDS.RDS",
            nuseds.pop.info ="DATA/LookupFiles/aopen-data-portaljanuary-2019conservation_unit_system_sites.csv",
           epad.rds.file =  "DATA/RDSFiles/EPAD.RDS")}
}
\keyword{matching/merging}
\keyword{record}
