% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{plotGrid}
\usage{
plotGrid(grid.plot, x.label = "X Label", y.label = "Y Label",
  plot.title = "Title")
}
\arguments{
\item{grid.plot}{dataframe with numerical values}

\item{x.label}{columns of the grid}

\item{y.label}{rows of the grid}

\item{plot.title}{title for the plot}
}
\description{
Function to plot a heatmap based on a dataframe/matrix
}
\keyword{heatmap}
