% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizeData}
\alias{summarizeData}
\title{summarizeData}
\usage{
summarizeData(
  nuseds.rds.file = "DATA/RDSFiles/nuSEDS_mod.RDS",
 
    nuseds.pop.info = "DATA/LookupFiles/aopen-data-portaljanuary-2019conservation_unit_system_sites.csv",
  epad.rds.file = "DATA/RDSFiles/EPAD_mod.RDS",
  mrp.rds.file = "DATA/RDSFiles/MRP_mod.RDS",
  mrp.rds.details.file = "DATA/RDSFiles/MRP_Details_mod.RDS",
  rosetta.cu = read.csv("DATA/LookupFiles/Generated_RosettaFile_CU.csv",
    stringsAsFactors = FALSE),
  rosetta.pop = read.csv("DATA/LookupFiles/Generated_RosettaFile_Pop.csv",
    stringsAsFactors = FALSE),
  cu.summary = read.csv("OUTPUT/Summary_CU.csv", stringsAsFactors = FALSE),
  out.folder = "OUTPUT/Summaries",
  tracking.folder = "DATA/TrackingFiles"
)
}
\arguments{
\item{nuseds.rds.file}{path to the modified nuSEDS input file (in RDS format).}

\item{nuseds.pop.info}{path to the pop/site info file from nuSEDS (in csv format)}

\item{epad.rds.file}{path to the modified EPAD input file (in RDS format)}

\item{mrp.rds.file}{path to the modified MRP summary input file (in RDS format)}

\item{out.folder}{path for storing output (CU summary stored there)}

\item{tracking.folder}{path for storing tracking files (site mismatches etc. stored there)}

\item{epad.rds.details.file}{path to the modified EPAD details input file (in RDS format)}
}
\description{
Function to generate various data summaries (number of records by DB, input for heatmap of tag recoveries)
Target folders are generated if they don't exist. Uses files generated with matchRecords().
}
\examples{
\dontrun{summarizeData( nuseds.rds.file = "DATA/RDSFiles/nuSEDS_mod.RDS",
        nuseds.pop.info = "DATA/LookupFiles/aopen-data-portaljanuary-2019conservation_unit_system_sites.csv",
        epad.rds.file =  "DATA/RDSFiles/EPAD_mod.RDS",
        mrp.rds.file =  "DATA/RDSFiles/MRP_mod.RDS",
        mrp.rds.details.file =  "DATA/RDSFiles/MRP_Details_mod.RDS",
        rosetta.cu = read.csv("DATA/LookupFiles/Generated_RosettaFile_CU.csv",stringsAsFactors = FALSE),
        rosetta.pop = read.csv("DATA/LookupFiles/Generated_RosettaFile_Pop.csv",stringsAsFactors = FALSE),
        cu.summary = read.csv("OUTPUT/Summary_CU.csv", stringsAsFactors =  FALSE), 
        out.folder = "OUTPUT/Summaries",
        tracking.folder = "DATA/TrackingFiles")
	}
}
\keyword{record}
\keyword{summaries}
