% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls2csv.R
\name{xls2csv}
\alias{xls2csv}
\title{xls2csv}
\usage{
xls2csv(
  file.in,
  file.out,
  db.type = "nuSEDS",
  sheet.read = 1,
  na.strings = c("", "NA")
)
}
\arguments{
\item{file.in}{path to the input file}

\item{file.out}{path to the output file}

\item{db.type}{one of "nuSEDS","EPAD", or "MRP"}

\item{sheet.read}{number specifying the tab to read in. Default is 1.}

\item{na.strings}{vector with strings to be interpreted as NA. Default is c("","NA")}
}
\description{
Function to convert xlsx files with salmon data to csv files. This function uses the {readxl} package which is part of the {tidyverse} package.
}
\examples{
\dontrun{xls2csv("test.xls","test.csv",db.type="nuSEDS")}
}
\keyword{conversion,}
\keyword{csv}
\keyword{file}
\keyword{xls,}
